[Fix]: 프로젝트 저장 메서드 변수명 변경 & API PATH 수정 (#26)
