+++ src/main/java/team/gdsc/code2cv/feature/project/service/ProjectService.java
@@ -1,17 +1,11 @@
 package team.gdsc.code2cv.feature.project.service;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import team.gdsc.code2cv.feature.project.domain.ProjectCommand;
 import team.gdsc.code2cv.feature.project.dto.ProjectRes;
 import team.gdsc.code2cv.feature.project.entity.Project;
 import team.gdsc.code2cv.feature.project.repository.GithubRepoApiResponse;
@@ -29,7 +23,7 @@ public class ProjectService {
 	private final ProjectUpsertUseCase projectUpsertUseCase;
 
 	@Transactional(readOnly = true)
-	public List<ProjectRes.ProjectDto> getAllProjects(Long userId) {
+	public List<ProjectRes.ProjectDto> getAllUserProjects(Long userId) {
 		List<Project> projects = projectRepository.findAllByUserId(userId);
 		return projects.stream()
 			// 수정일 최신순 정렬
@@ -43,50 +37,10 @@ public class ProjectService {
 	 * 1. 사용자의 깃토큰으로 Github Repository 정보를 가져온다.
 	 * 2. 가져온 정보를 프로젝트로 변환하여 저장하는 과정을 트랜잭션을 위해 [UseCase]에 위임한다.
 	 */
-	public void analyzeProject(Long userId) {
+	public void saveUserProjects(Long userId) {
 		User user = userRepository.findByIdOrThrow(userId);
 		String githubToken = user.getGithubAccessToken();
 		List<GithubRepoApiResponse> repositoriesInfo = githubRepoClient.getAllRepositoriesInfo(githubToken);
 		projectUpsertUseCase.invoke(repositoriesInfo, userId);
 	}
-}
-
-@Component
-@RequiredArgsConstructor
-class ProjectUpsertUseCase {
-	private final ProjectRepository projectRepository;
-	private final UserRepository userRepository;
-
-	/**
-	 * Github Repository 정보를 프로젝트로 변환하여 저장한다.
-	 * 1. 사용자의 프로젝트 목록을 조회한다.
-	 * 2. 존재한다면 업데이트, 존재하지 않는다면 생성한다.
-	 */
-	@Transactional
-	public void invoke(List<GithubRepoApiResponse> repositoriesInfo, Long userId) {
-		User user = userRepository.getReferenceById(userId);
-		List<Project> projects = projectRepository.findAllByUserId(userId);
-
-		List<ProjectCommand.Create> createCommands = new ArrayList<>();
-		Map<Project, ProjectCommand.Update> updateCommands = new HashMap<>();
-
-		for (GithubRepoApiResponse repositoryInfo : repositoriesInfo) {
-			Optional<Project> project = findProject(repositoryInfo, projects);
-			project.ifPresentOrElse(
-				p -> updateCommands.put(p, repositoryInfo.toUpdateCommand()),
-				() -> createCommands.add(repositoryInfo.toCreateCommand())
-			);
-		}
-		createCommands.stream()
-			.map(c -> Project.create(c, user))
-			.forEach(projectRepository::save);
-		updateCommands.forEach(Project::update);
-	}
-
-	private Optional<Project> findProject(GithubRepoApiResponse repositoryInfo, List<Project> projects) {
-		return projects.stream()
-			.filter(p -> p.getRepoId().equals(repositoryInfo.id()))
-			.findAny();
-	}
-
 }
